Refactor deleting heartbeat topics
