+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/PulsarService.java
@@ -22,7 +22,6 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import static org.apache.commons.lang3.StringUtils.isBlank;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 import static org.apache.pulsar.broker.resourcegroup.ResourceUsageTransportManager.DISABLE_RESOURCE_USAGE_TRANSPORT_MANAGER;
-import static org.apache.pulsar.broker.service.HealthChecker.getHeartbeatTopicName;
 import static org.apache.pulsar.common.naming.SystemTopicNames.isTransactionInternalName;
 import com.google.common.annotations.VisibleForTesting;
 import io.netty.channel.ChannelInitializer;
@@ -75,7 +74,6 @@ import org.apache.bookkeeper.conf.ClientConfiguration;
 import org.apache.bookkeeper.mledger.LedgerOffloader;
 import org.apache.bookkeeper.mledger.LedgerOffloaderFactory;
 import org.apache.bookkeeper.mledger.LedgerOffloaderStats;
-import org.apache.bookkeeper.mledger.ManagedLedgerException;
 import org.apache.bookkeeper.mledger.ManagedLedgerFactory;
 import org.apache.bookkeeper.mledger.impl.NullLedgerOffloader;
 import org.apache.bookkeeper.mledger.offload.Offloaders;
@@ -450,41 +448,6 @@ public class PulsarService implements AutoCloseable, ShutdownService {
         }
     }
 
-    private boolean isManagedLedgerNotFoundException(Throwable e) {
-        Throwable realCause = e.getCause();
-        return realCause instanceof ManagedLedgerException.MetadataNotFoundException
-                || realCause instanceof MetadataStoreException.NotFoundException;
-    }
-
-    private void deleteHeartbeatResource() {
-        if (this.brokerService != null) {
-            LOG.info("forcefully delete heartbeat topic when close broker");
-
-            String heartbeatTopicNameV1 = getHeartbeatTopicName(getBrokerId(), getConfiguration(), false);
-            String heartbeatTopicNameV2 = getHeartbeatTopicName(getBrokerId(), getConfiguration(), true);
-
-            try {
-                this.brokerService.deleteTopic(heartbeatTopicNameV1, true).get();
-            } catch (Exception e) {
-                if (!isManagedLedgerNotFoundException(e)) {
-                    LOG.error("Closed with errors in delete heartbeat topic [{}]",
-                            heartbeatTopicNameV1, e);
-                }
-            }
-
-            try {
-                this.brokerService.deleteTopic(heartbeatTopicNameV2, true).get();
-            } catch (Exception e) {
-                if (!isManagedLedgerNotFoundException(e)) {
-                    LOG.error("Closed with errors in delete heartbeat topic [{}]",
-                            heartbeatTopicNameV2, e);
-                }
-            }
-
-            LOG.info("finish forcefully delete heartbeat topic when close broker");
-        }
-    }
-
     @Override
     public void close() throws PulsarServerException {
         try {
@@ -533,11 +496,6 @@ public class PulsarService implements AutoCloseable, ShutdownService {
             }
             // It only tells the Pulsar clients that this service is not ready to serve for the lookup requests
             state = State.Closing;
-
-            if (brokerId != null) {
-                // forcefully delete heartbeat topic when close broker
-                deleteHeartbeatResource();
-            }
 
             if (healthChecker != null) {
                 healthChecker.close();
